refactor(auth): simplify AuthAndUserInitializer by utilizing useAuthGuard for authentication checks
